requirements.txt
streamlit
google-generativeai
faiss-cpu
PyPDF2
python-docx
python-pptx
pandas
numpy
python-dotenv
sentence-transformers
langchain-huggingface
